requirements.txt
asyncio==3.4.3
discord.py==1.6.0
http3==0.6.7
